+++ src/main/java/com/github/admissionCommittee/model/SchoolCertificate.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
+import lombok.ToString;
 
 import javax.persistence.AttributeOverride;
 import javax.persistence.CollectionTable;
@@ -24,6 +25,7 @@ import java.util.Map;
 @Table(name = "school_certificate_id")
 @EqualsAndHashCode(callSuper = true, doNotUseGetters = true)
 @AttributeOverride(name = "id", column = @Column(name = "school_certificate_id", nullable = false))
+@ToString(exclude = {"user"})
 public class SchoolCertificate extends AbstractEntity {
 
     @OneToOne
