+++ src/main/java/com/github/admissionCommittee/model/Sheet.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
+import lombok.ToString;
 
 import javax.persistence.AttributeOverride;
 import javax.persistence.Column;
@@ -19,6 +20,7 @@ import javax.persistence.Table;
 @Table(name = "sheet")
 @EqualsAndHashCode(callSuper = true, doNotUseGetters = true)
 @AttributeOverride(name = "id", column = @Column(name = "sheet_id", nullable = false))
+@ToString(exclude = "user")
 public class Sheet extends AbstractEntity {
 
     @OneToOne
