--- src/main/java/com/github/admissionCommittee/model/UserTypeEnum.java
@@ -1,6 +1,0 @@
-package com.github.admissionCommittee.model;
-
-public enum UserTypeEnum {
-    USER,
-    ADMIN
-}
