+++ src/main/java/com/github/admissionCommittee/model/enums/UserTypeEnum.java
@@ -1,0 +1,6 @@
+package com.github.admissionCommittee.model.enums;
+
+public enum UserTypeEnum {
+    USER,
+    ADMIN
+}
